+++ javascript/src/formatting/formatForStatus.ts
@@ -1,0 +1,13 @@
+import { TestStepResultStatus } from '@cucumber/messages'
+
+import { TextBuilder } from '../TextBuilder'
+import { Theme } from '../types'
+
+export function formatForStatus(
+  status: TestStepResultStatus,
+  text: string,
+  theme: Theme,
+  stream: NodeJS.WritableStream
+): string {
+  return new TextBuilder(stream).append(text).build(theme.status?.all?.[status])
+}