+++ javascript/src/formatting/formatStatusName.ts
@@ -1,0 +1,5 @@
+import { TestStepResultStatus } from '@cucumber/messages'
+
+export function formatStatusName(status: TestStepResultStatus) {
+  return `${status.charAt(0).toUpperCase() + status.slice(1).toLowerCase()}`
+}