+++ javascript/src/formatting/formatUndefinedParameterType.ts
@@ -1,0 +1,5 @@
+import { UndefinedParameterType } from '@cucumber/messages'
+
+export function formatUndefinedParameterType(upt: UndefinedParameterType): string {
+  return `'${upt.name}' in '${upt.expression}'`
+}