+++ javascript/src/queries/findAllSuggestions.ts
@@ -1,0 +1,10 @@
+import { Suggestion } from '@cucumber/messages'
+import { Query } from '@cucumber/query'
+
+export function findAllSuggestions(query: Query): ReadonlyArray<Suggestion> {
+  return query
+    .findAllTestCaseFinished()
+    .map((testCaseFinished) => query.findPickleBy(testCaseFinished))
+    .filter((pickle) => !!pickle)
+    .flatMap((pickle) => query.findSuggestionsBy(pickle))
+}